requirements.txt
flask~=3.1.0
numpy~=2.1.3
scikit-learn~=1.5.2
scipy
matplotlib
Werkzeug~=3.1.3
pandas~=2.2.3
torch~=2.5.1
pillow~=11.0.0
open_clip_torch~=2.29.0